+++ src/app/cart/index.ts
@@ -1,2 +1,3 @@
 export * from './components/cart-list/cart-list.component';
+export * from './components/cart-item/cart-item.component';
 export * from './services/cart.service';