+++ src/app/cart/services/cart.service.ts
@@ -1,34 +1,53 @@
 import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
+
+import { ProductCartModel } from '../../products/models/products.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CartService {
+  public itemsChanged$: Subject<any> = new Subject<any>();
+  private purchasedItems: Array<ProductCartModel> = [];
 
   constructor() { }
 
-  // вместо object можно использовать интерфейс построенный на базе ProductModel,
-  // расширив его необходимыми полями
-  getProducts(): object[] {
-    return [
-      {
-        id: 1,
-        name: 'Water',
-        price: 2,
-        amount: 5,
-      },
-      {
-        id: 2,
-        name: 'Laptop',
-        price: 1500,
-        amount: 1,
-      },
-      {
-        id: 3,
-        name: 'Sneakers',
-        price: 200,
-        amount: 1,
-      },
-    ];
+  getProducts(): ProductCartModel[] {
+    return [...this.purchasedItems];
+  }
+
+  addProduct(item: ProductCartModel): void {
+    const exist = this.purchasedItems.findIndex(i => i.id === item.id);
+    if (exist !== -1) {
+      this.purchasedItems[exist].quantity += 1;
+      this.calcTotalPrice(exist);
+    } else {
+      item.quantity = 1;
+      item.totalPrice = item.price;
+      this.purchasedItems.push(item);
+    }
+    this.itemsChanged$.next();
+  }
+
+  deleteProduct(id: number): void {
+    this.purchasedItems = this.purchasedItems.filter(i => i.id !== id);
+    this.itemsChanged$.next();
+  }
+
+  updateProduct(id: number, key: string, value: any): void {
+    const exist = this.purchasedItems.findIndex(i => i.id === id);
+    if (exist !== -1) {
+      this.purchasedItems[exist][key] = value;
+      this.calcTotalPrice(exist);
+      this.itemsChanged$.next();
+    }
+  }
+
+  getSummaryBy(key: string): number {
+    return this.purchasedItems.reduce((prev, curr) => prev + curr[key], 0);
+  }
+
+  private calcTotalPrice(index: number): void {
+    this.purchasedItems[index].totalPrice = this.purchasedItems[index].quantity * this.purchasedItems[index].price;
   }
 }