+++ src/app/products/components/product/product.component.ts
@@ -1,11 +1,17 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+
+import { Product } from '../../models/products.model';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
-  styleUrls: ['./product.component.scss']
+  styleUrls: ['./product.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductComponent implements OnInit {
+  @Input() product: Product;
+
+  @Output() buy: EventEmitter<number> = new EventEmitter<number>();
 
   constructor() { }
 
@@ -13,6 +19,6 @@ export class ProductComponent implements OnInit {
   }
 
   onBuy(): void {
-    console.log('The item was bought!');
+    this.buy.emit(this.product.id);
   }
 }