+++ src/app/products/products.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ProductComponent, ProductListComponent } from './components';
+import { SharedModule } from '../shared/shared.module';
+
+
+
+@NgModule({
+  declarations: [
+    ProductListComponent,
+    ProductComponent,
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+  ],
+  exports: [
+    ProductListComponent,
+  ],
+})
+export class ProductsModule { }